+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Quiqr App CHANGELOG
 ## **WORK IN PROGRESS**
+- fix: #452 make config.json source path agnostic
+- feature: copy collection item to lang
 
 ## 0.18.9 (2024-09-27)
 - feature: AI Assistent for text fields. Using ChatGPT from OpenAI.